Update mocha and use setup global beforeEach/afterEach
