+++ spec/javascripts/bindEntityEvents.spec.js
@@ -1,7 +1,5 @@
 describe('Marionette.bindEntityEvents', function() {
   'use strict';
-  beforeEach(global.setup);
-  afterEach(global.teardown);
 
   beforeEach(function() {
     this.target = {
