+++ spec/javascripts/collectionView.attachToDOM.spec.js
@@ -1,7 +1,5 @@
 describe('attaching collection view to existing DOM structure', function() {
   'use strict';
-  beforeEach(global.setup);
-  afterEach(global.teardown);
 
   describe('when initializing a collection view', function() {
     beforeEach(function() {
