+++ spec/javascripts/collectionView.emptyView.spec.js
@@ -1,7 +1,5 @@
 describe('collectionview - emptyView', function() {
   'use strict';
-  beforeEach(global.setup);
-  afterEach(global.teardown);
 
   beforeEach(function() {
     this.ItemView = Backbone.Marionette.ItemView.extend({
