+++ spec/javascripts/collectionView.spec.js
@@ -1,7 +1,5 @@
 describe('collection view', function() {
   'use strict';
-  beforeEach(global.setup);
-  afterEach(global.teardown);
 
   beforeEach(function() {
     // Shared View Definitions
