+++ spec/javascripts/commands.spec.js
@@ -1,7 +1,5 @@
 describe('application commands', function() {
   'use strict';
-  beforeEach(global.setup);
-  afterEach(global.teardown);
 
   describe('when creating an instance of an Application', function() {
     beforeEach(function() {
