+++ spec/javascripts/compositeView-childViewContainer.spec.js
@@ -1,7 +1,5 @@
 describe('composite view - childViewContainer', function() {
   'use strict';
-  beforeEach(global.setup);
-  afterEach(global.teardown);
 
   beforeEach(function() {
     this.Model = Backbone.Model.extend({});
