+++ spec/javascripts/compositeView.spec.js
@@ -1,7 +1,5 @@
 describe('composite view', function() {
   'use strict';
-  beforeEach(global.setup);
-  afterEach(global.teardown);
 
   beforeEach(function() {
     var suite = this;
