+++ spec/javascripts/layoutView.spec.js
@@ -1,7 +1,5 @@
 describe('layoutView', function() {
   'use strict';
-  beforeEach(global.setup);
-  afterEach(global.teardown);
 
   beforeEach(function() {
     this.template = function() {
