+++ spec/javascripts/normalizeMethods.spec.js
@@ -1,7 +1,5 @@
 describe('normalizeMethods', function() {
   'use strict';
-  beforeEach(global.setup);
-  afterEach(global.teardown);
 
   describe('when normalizeMethods is called with a hash of functions and strings', function() {
     beforeEach(function() {
