+++ spec/javascripts/unbindEntityEvents.spec.js
@@ -1,7 +1,5 @@
 describe('Marionette.unbindEntityEvents', function() {
   'use strict';
-  beforeEach(global.setup);
-  afterEach(global.teardown);
 
   beforeEach(function() {
     this.target = {
