+++ spec/javascripts/view.uiEventAndTriggers.spec.js
@@ -1,7 +1,5 @@
 describe('view ui event trigger configuration', function() {
   'use strict';
-  beforeEach(global.setup);
-  afterEach(global.teardown);
 
   describe('@ui syntax within events and triggers', function() {
     beforeEach(function() {
